MVPs: Trim leading @ from social usernames
Some users have kept the @ in front of the username, so there are actually two @@ in front of their username. This commit will trim the username to make sure there is only a single @.

This appears to be isolated to the MVPs list as the profile page doesn't suffer from it.