Image gallery overlay not vertically centered #429
This is a fix for the issue #429 (https://github.com/umbraco/OurUmbraco/issues/429)
Unfortunatly there was no real way for me to test this as i can preview content thats in the link, but based on the information in the ticket it should work fine.

I also took the liberty of tidying up the _package-detail file.